refactor: remove dead code and consolidate duplicated helpers
- Delete deprecated resolveCritDir and its test
- Replace _ = err no-op with clear comment in auth.go
- Make fileHash delegate to computeFileHash instead of duplicating SHA256
- Replace inline web-comment merge in runFetch with mergeWebComments call
- Move highestWebIndex to share.go near its only consumer
